Get rid of floats.

+++ headers/character.h
@@ -11,13 +11,13 @@ struct Character {
     int cx, cy;
     int speed, move_speed, speed_relative;
     
-    float grav;
-    float grav_fall;
-    float grav_add;
-    float grav_max;
-    float grav_min;
-    float grav_zero;
-    float grav_relative;
+    double grav;
+    double grav_fall;
+    double grav_add;
+    double grav_max;
+    double grav_min;
+    double grav_zero;
+    int grav_relative;
     
     bool on_ground;
     bool on_platform;
@@ -34,7 +34,7 @@ struct Character *chara_create(struct Map *map, const int x, const int y,
                                                 
 void chara_free(struct Character *chara);
 
-void chara_set_gravity(struct Character *chara, float add, float min, float max);
+void chara_set_gravity(struct Character *chara, double add, double min, double max);
 void chara_update(struct Character *chara, const bool has_control,
                                            const bool fall_off);
